requirements.txt
# TRUST: Leveraging Text Robustness for Unsupervised Domain Adaptation
# Python >= 3.8 required

# PyTorch (install with CUDA support for your system)
# Visit https://pytorch.org/get-started/locally/ for installation instructions
torch>=1.12.0
torchvision>=0.13.0

# Transformers (Hugging Face)
transformers>=4.25.0

# Image processing
Pillow>=9.0.0

# Scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Logging and experiment tracking
wandb>=0.13.0
tqdm>=4.64.0

# Text processing
yake>=0.4.8
rake-nltk>=1.0.6
spacy>=3.4.0

# NLP utilities
nltk>=3.7
